+++ packages/next/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @collie-lang/next
 
+## 2.0.1
+
+### Patch Changes
+
+- Version 5
+- Updated dependencies
+  - @collie-lang/compiler@6.0.1
+  - @collie-lang/webpack@5.0.1
+
+## 2.0.0
+
+### Major Changes
+
+- 7efa70c: Updated compiler and CLI
+
+### Patch Changes
+
+- Updated dependencies [7efa70c]
+  - @collie-lang/compiler@6.0.0
+  - @collie-lang/webpack@5.0.0
+
 ## 1.0.3
 
 ### Patch Changes